Re-arrange authorization API.
