+++ test/server/requests/logout.rest
@@ -1,3 +1,3 @@
-DELETE http://localhost:4200/api/logout
+DELETE http://localhost:4200/api/session
 Content-type: application/json
-Cookie: token=YWRtaW49MmNlMWFkMDgtMzY4Ni00Y2EwLTliMzEtYzQxYTRjZjhjMDFl; Path=/
+Cookie: token=YWRtaW49NGI4ZGRiNDQtZjNmYy00MjhiLThiNjYtNmE1OWQyZmU5NzE1; Path=/
