+++ test/server/requests/revoke.rest
@@ -1,4 +1,4 @@
-POST http://localhost:4200/api/revoke-all
+POST http://localhost:4200/api/session/revoke-all
 Content-type: application/json
 
 {
